+++ Tests/SentryTests/Networking/SentryReachabilityTests.m
@@ -1,5 +1,11 @@
+#import "SentryReachability+Private.h"
 #import "SentryReachability.h"
 #import <XCTest/XCTest.h>
+
+@interface TestSentryReachabilityObserver : NSObject <SentryReachabilityObserver>
+@end
+@implementation TestSentryReachabilityObserver
+@end
 
 #if !TARGET_OS_WATCH
 @interface SentryConnectivityTest : XCTestCase
@@ -20,30 +26,62 @@
 
 - (void)testConnectivityRepresentations
 {
-    XCTAssertEqualObjects(@"none", SentryConnectivityFlagRepresentation(0));
-    XCTAssertEqualObjects(
-        @"none", SentryConnectivityFlagRepresentation(kSCNetworkReachabilityFlagsIsDirect));
+    XCTAssertEqualObjects(SentryConnectivityNone, SentryConnectivityFlagRepresentation(0));
+    XCTAssertEqualObjects(SentryConnectivityNone,
+        SentryConnectivityFlagRepresentation(kSCNetworkReachabilityFlagsIsDirect));
 #    if SENTRY_HAS_UIKIT
     // kSCNetworkReachabilityFlagsIsWWAN does not exist on macOS
-    XCTAssertEqualObjects(
-        @"none", SentryConnectivityFlagRepresentation(kSCNetworkReachabilityFlagsIsWWAN));
-    XCTAssertEqualObjects(@"cellular",
+    XCTAssertEqualObjects(SentryConnectivityNone,
+        SentryConnectivityFlagRepresentation(kSCNetworkReachabilityFlagsIsWWAN));
+    XCTAssertEqualObjects(SentryConnectivityCellular,
         SentryConnectivityFlagRepresentation(
             kSCNetworkReachabilityFlagsIsWWAN | kSCNetworkReachabilityFlagsReachable));
 #    endif // SENTRY_HAS_UIKIT
-    XCTAssertEqualObjects(
-        @"wifi", SentryConnectivityFlagRepresentation(kSCNetworkReachabilityFlagsReachable));
-    XCTAssertEqualObjects(@"wifi",
+    XCTAssertEqualObjects(SentryConnectivityWiFi,
+        SentryConnectivityFlagRepresentation(kSCNetworkReachabilityFlagsReachable));
+    XCTAssertEqualObjects(SentryConnectivityWiFi,
         SentryConnectivityFlagRepresentation(
             kSCNetworkReachabilityFlagsReachable | kSCNetworkReachabilityFlagsIsDirect));
 }
 
-- (void)testUniqueKeyForInstances
+- (void)testMultipleReachabilityObservers
 {
-    SentryReachability *anotherReachability = [[SentryReachability alloc] init];
-    XCTAssertNotEqualObjects(
-        [self.reachability keyForInstance], [anotherReachability keyForInstance]);
-    XCTAssertEqualObjects([self.reachability keyForInstance], [self.reachability keyForInstance]);
+    SentryReachability *reachability = [[SentryReachability alloc] init];
+
+    XCTestExpectation *aExp =
+        [self expectationWithDescription:
+                  @"reachability state change for observer monitoring https://sentry.io"];
+    aExp.expectedFulfillmentCount = 5;
+    TestSentryReachabilityObserver *a = [[TestSentryReachabilityObserver alloc] init];
+    [reachability addObserver:a
+                 withCallback:^(__unused BOOL connected,
+                     NSString *_Nonnull __unused typeDescription) { [aExp fulfill]; }];
+
+    SentryConnectivityCallback(reachability.sentry_reachability_ref,
+        kSCNetworkReachabilityFlagsReachable, nil); // ignored, as it's the first callback
+    SentryConnectivityCallback(
+        reachability.sentry_reachability_ref, kSCNetworkReachabilityFlagsInterventionRequired, nil);
+
+    XCTestExpectation *bExp =
+        [self expectationWithDescription:
+                  @"reachability state change for observer monitoring https://google.io"];
+    bExp.expectedFulfillmentCount = 2;
+    TestSentryReachabilityObserver *b = [[TestSentryReachabilityObserver alloc] init];
+    [reachability addObserver:b
+                 withCallback:^(__unused BOOL connected,
+                     NSString *_Nonnull __unused typeDescription) { [bExp fulfill]; }];
+
+    SentryConnectivityCallback(
+        reachability.sentry_reachability_ref, kSCNetworkReachabilityFlagsReachable, nil);
+    SentryConnectivityCallback(
+        reachability.sentry_reachability_ref, kSCNetworkReachabilityFlagsInterventionRequired, nil);
+
+    [reachability removeObserver:b];
+
+    SentryConnectivityCallback(
+        reachability.sentry_reachability_ref, kSCNetworkReachabilityFlagsReachable, nil);
+
+    [self waitForExpectationsWithTimeout:1.f handler:nil];
 }
 
 @end